requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Booktype==1.5
click==8.1.8
Django==5.1.6
fastapi==0.115.8
fastapi-sessions==0.3.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
simplejson==3.20.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
